prep for creating the nodemap window
