+++ Assets/Merino/Editor/Scripts/MerinoNodemapWindow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4079d70fef2071c4d9fc01ea09e2edfb
+guid: d7382f9e802b52e49bb076382ad82e07
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
