+++ Assets/Merino/Editor/Scripts/MerinoYarnEditorWindow.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 4079d70fef2071c4d9fc01ea09e2edfb
+guid: c3cf55b48c197bb40a09696dead1d7c2
+timeCreated: 1467106475
+licenseType: Pro
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
